// src/client/toteco/models/publication.ts
/* tslint:disable */
/* eslint-disable */
/**
 * TOTECO SERVICE
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * Contact: ire.cunba@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import { Establishment } from './establishment';
import { User } from './user';
 /**
 * 
 *
 * @export
 * @interface Publication
 */
export interface Publication {

    /**
     * @type {string}
     * @memberof Publication
     */
    id?: string;

    /**
     * @type {number}
     * @memberof Publication
     */
    created?: number;

    /**
     * @type {number}
     * @memberof Publication
     */
    updated?: number;

    /**
     * @type {number}
     * @memberof Publication
     */
    totalPrice?: number;

    /**
     * @type {number}
     * @memberof Publication
     */
    totalScore?: number;

    /**
     * @type {string}
     * @memberof Publication
     */
    photo?: string;

    /**
     * @type {string}
     * @memberof Publication
     */
    comment?: string;

    /**
     * @type {Establishment}
     * @memberof Publication
     */
    establishment?: Establishment;

    /**
     * @type {User}
     * @memberof Publication
     */
    user?: User;
}