// src/client/toteco/models/publication-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * TOTECO SERVICE
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * Contact: ire.cunba@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

 /**
 * 
 *
 * @export
 * @interface PublicationDTO
 */
export interface PublicationDTO {

    /**
     * @type {number}
     * @memberof PublicationDTO
     */
    totalPrice?: number;

    /**
     * @type {number}
     * @memberof PublicationDTO
     */
    totalScore?: number;

    /**
     * @type {string}
     * @memberof PublicationDTO
     */
    photo?: string;

    /**
     * @type {string}
     * @memberof PublicationDTO
     */
    comment?: string;

    /**
     * @type {string}
     * @memberof PublicationDTO
     */
    establishmentId?: string;

    /**
     * @type {string}
     * @memberof PublicationDTO
     */
    userId?: string;
}